return (node, index) successor from remove*
previously, remove* returned the key of the (ascending-only) successor
to the removed element. now the exact successor (ascending or descending)
is returned, removing the need for a follow-up traversal when the tree
structure changes.

improvements to the submap iterators mean that getBoundaryNode() is no
longer required when iterating, just when removing elements.

this also fixes maps that allowed the null key, which couldn't distinguish
a null-keyed successor from no successor.